Support multi-line strings in descriptions
When using line-breaks in type/fields descriptions or in a deprecation reason,
an invalid SLD schema would be generated, because a newline character would be
inserted when replacing the non-directive definitions with the directive definitions.

To prevent that regex converts `\n` to a newline character, we are escaping the
backslashes.